:art: move cut_sequence_chunks to utils.py
not meant to be publicly exposed
